--- a/src/inline/mod.rs
+++ b/src/inline/mod.rs
@@ -1,6 +1,8 @@
 use std::mem::MaybeUninit;
 
+/// Inline HashMap
 pub mod hashmap;
+/// Inline vector
 pub mod vec;
 
 fn uninit_array<T, const N: usize>() -> [MaybeUninit<T>; N] {