--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1,45 +1,8 @@
-//! tl is an efficient and easy to use HTML parser written in Rust.
-//!
-//! It does minimal to no copying during parsing by borrowing parts of the input string.
-//! Additionally, it keeps track of parsed elements and stores elements with an id attribute
-//! in an internal HashMap, which makes element lookups by ID/class name very fast.
-//!
-//! ## Examples
-//! Finding an element by its id attribute and printing the inner text:
-//! ```rust
-//! let input = r#"<p id="text">Hello</p>"#;
-//! let dom = tl::parse(input);
-//!
-//! let element = dom.get_element_by_id("text").expect("Failed to find element");
-//!
-//! println!("Inner text: {}", element.inner_text());
-//! ```
-//!
-//! ## Owned DOM
-//! Calling `tl::parse()` returns a DOM struct that borrows from the input string, which means that the string must be kept alive.
-//! The input string must outlive this DOM. If this is not acceptable or you need to keep the DOM around for longer,
-//! consider using `tl::parse_owned()`.
-//! `VDomGuard` takes ownership over the string, which means you don't have to keep the string around.
-//! ```rust
-//! // Notice how it takes ownership over the string:
-//! let dom_guard = unsafe { tl::parse_owned(String::from(r#"<p id="text">Hello</p>"#)) };
-//!
-//! // Obtain reference to underlying VDom
-//! let dom = dom_guard.get_ref();
-//!
-//! // Now, use `dom` as you would if it was a regular `VDom`
-//! let element = dom.get_element_by_id("text").expect("Failed to find element");
-//!
-//! println!("Inner text: {}", element.inner_text());
-//! ```
-//!
-//! ## Bytes
-//! Some methods return a `Bytes` struct, which is an internal struct that is used to borrow
-//! a part of the input string. This is mainly used over a raw `&[u8]` for its `Debug` implementation.
-
+#![doc = include_str!("../README.md")]
 #![deny(missing_docs)]
 
 mod bytes;
+mod inline;
 mod parser;
 mod stream;
 #[cfg(test)]
@@ -48,8 +11,12 @@ mod util;
 mod vdom;
 
 pub use bytes::{AsBytes, Bytes};
-use parser::Parser;
-pub use parser::{tag::Attributes, tag::HTMLTag, tag::Node, HTMLVersion, Tree};
+pub use inline::{
+    hashmap::{InlineHashMap, InlineHashMapIterator},
+    vec::{InlineVec, InlineVecIter},
+};
+pub use parser::Parser;
+pub use parser::{handle::NodeHandle, tag::Attributes, tag::HTMLTag, tag::Node, HTMLVersion, Tree};
 pub use vdom::{VDom, VDomGuard};
 
 /// Parses the given input string