--- a/src/parser/mod.rs
+++ b/src/parser/mod.rs
@@ -1,5 +1,6 @@
 pub mod base;
 pub mod constants;
+pub mod handle;
 pub mod tag;
 
 pub use base::*;