- avoids Rc allocations
- inline vectors to avoid allocating for small collections
- fixes #7
- disable id/class tracking by default

before:
```
tl-throughput/tl        time:   [1.4386 ms 1.4425 ms 1.4461 ms]
                        thrpt:  [216.29 MiB/s 216.84 MiB/s 217.42 MiB/s]
```

after:
```
tl-throughput/tl        time:   [743.77 us 744.32 us 744.88 us]
                        thrpt:  [419.91 MiB/s 420.22 MiB/s 420.53 MiB/s]
```